Additional changes
Wrote the codes of 'Ekle' button click action to add new user record to
database.
Added a new case to showMessage method in Helper class(to notify that
operation is successful)
Created a new method to update table content(more like refreshin it)
when we inserted a new doctor record.